# data-service/backend/api/data/v1/openapi.yaml
openapi: 3.0.4
info:
  title: Tidewise Data Service API
  version: 1.0.0
  description: >-
    Data Service 内部 HTTP REST + JSON 合同。消费方拥有小型手写 typed client，
    并以本文件的 operationId、schema drift anchor 和静态合同测试防止漂移。
...
servers:
  - url: /
    description: 由运行环境注入 Data Service origin；合同只冻结路径。
security:
  - ServiceBearer: []
...
paths:
...
  /api/data/v1/research/themes:
    get:
      tags: [Research]
      operationId: listResearchThemes
      x-client-drift-anchor: data.v1.listResearchThemes
      x-required-service-scope: data.research.read
      x-retry-policy: safe-get
      x-timeout-budget-ms: 5000
      summary: 获取 Miniapp Research theme 页面聚合
      description: >-
        返回半开区间 [published_from, published_to) 内成功发布的 Theme Aggregate，按
        published_at DESC、id ASC 稳定排序并通过 cursor 分页。published_from 和
        published_to 必须成对传入，与兼容参数 window_hours 互斥。
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/WindowHours'
        - $ref: '#/components/parameters/PublishedFrom'
        - $ref: '#/components/parameters/PublishedTo'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Theme 页面聚合
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchThemeListEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/data/v1/research/themes/{theme_id}:
    parameters:
      - $ref: '#/components/parameters/ThemeID'
    get:
      tags: [Research]
      operationId: getResearchTheme
      x-client-drift-anchor: data.v1.getResearchTheme
      x-required-service-scope: data.research.read
      x-retry-policy: safe-get
      x-timeout-budget-ms: 5000
      summary: 获取一个 Research theme 及其证据聚合
      description: 按 theme_id 读取已发布主题，不受列表时间窗口限制。
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/WindowHours'
      responses:
        '200':
          description: Theme 详情聚合
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchThemeDetailEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/data/v1/research-analysis-context:
    get:
      tags: [Research Analysis]
      operationId: listResearchAnalysisContext
      x-client-drift-anchor: data.v1.listResearchAnalysisContext
      x-required-service-scope: data.research.read
      x-retry-policy: safe-get-restart-from-first-page-on-cursor-mismatch
      x-timeout-budget-ms: 15000
      summary: 按显式研究时间窗口读取正式 Event Semantics 批量上下文
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - name: discovery_window_start
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/UTCTimestamp'
        - name: discovery_window_end
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/UTCTimestamp'
        - name: analysis_as_of
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/UTCTimestamp'
        - name: prediction_horizon_start
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UTCTimestamp'
        - name: prediction_horizon_end
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UTCTimestamp'
        - name: page_size
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 50
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: 实时正式 Event Semantic Bundles 与解析字典
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchAnalysisContextEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/ResearchAnalysisContextInconsistent'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/ResearchResourceLimit'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/data/v1/research-graph:search:
    post:
      tags: [Research Analysis]
      operationId: searchResearchGraph
      x-client-drift-anchor: data.v1.searchResearchGraph
      x-required-service-scope: data.research.read
      x-retry-policy: safe-idempotent-read-reduce-scope-on-resource-limit
      x-timeout-budget-ms: 15000
      summary: 按 Codex 显式范围执行受预算约束的确定性研究图谱检索
      description: >-
        Data 只执行确定性、只读、受预算约束的图遍历。industry_chain_entity_id 仅约束
        Industry Chain Graph Edge；全局 EntityRelation 仍由显式 relation_filters 控制。
        超过预算时整次失败，不返回静默截断的部分子图。
      parameters:
        - $ref: '#/components/parameters/RequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchGraphSearchRequest'
      responses:
        '200':
          description: 引用完整、稳定排序的可达子图
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchGraphSearchEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/ResearchResourceLimit'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/data/v1/research-theme-imports:
    post:
      tags: [Research Theme publication]
      operationId: publishResearchTheme
      x-client-drift-anchor: data.v1.publishResearchTheme
      x-required-service-scope: data.research.import
      x-retry-policy: idempotent-with-analysis-batch-id
      x-timeout-budget-ms: 15000
      x-canonicalization: rfc8785-sha256
      x-atomicity: one-theme-and-all-reason-trees-single-postgresql-transaction
      x-receipt-schema: research_theme_import_receipts
      x-transaction-membership: research-theme-import-receipt,research-theme,research-theme-impact,research-theme-event,research-reasoning-tree-import-receipt,research-reasoning-tree,research-reasoning-tree-event,research-reasoning-tree-node,research-reasoning-tree-node-signal
      summary: 原子发布一条不可变 Theme 及其全部 Reason Trees
      description: >-
        analysis_batch_id 是本次单 Theme Aggregate 的幂等身份。V2 formal
        分支保持正式本体引用与血缘校验；V3 analyst_snapshot 分支是分析师提交的
        不可变展示报告，只要求 Event 存在，Evidence 仅在提交时校验归属，不要求
        Entity、IndustryChain、VariableSignal 或 Graph Relation。Theme 与 1..N
        Reason Trees 在同一事务内全部发布或全部回滚。
      parameters:
        - $ref: '#/components/parameters/RequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              maxLength: 1048576
              x-max-body-bytes: 1048576
              oneOf:
                - $ref: '#/components/schemas/ResearchThemeImportRequest'
                - $ref: '#/components/schemas/ResearchThemeSnapshotImportRequest'
      responses:
        '200':
          description: 相同发布主体、analysis_batch_id 与 payload 的首次结果重放
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchThemeImportEnvelope'
        '201':
          description: Theme Aggregate 和 immutable receipt 已原子提交
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchThemeImportEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    ServiceBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque service identity token
      description: >-
        环境 secret 注入的 service identity。服务端从认证 principal 派生 caller identity
        与 scope；request body 不得声明或覆盖 caller identity。
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: 可选 caller request id；缺失时服务端生成。不得作为业务幂等 identity。
      schema:
        type: string
        minLength: 1
        maxLength: 128
    WindowHours:
      name: window_hours
      in: query
      required: false
      deprecated: true
      description: 兼容旧列表调用；不得与 published_from/published_to 同时传入。详情路由仅校验参数。
      schema:
        type: integer
        minimum: 1
        maximum: 168
        default: 24
    PublishedFrom:
      name: published_from
      in: query
      required: false
      description: 发布时间半开区间下界，必须为 UTC RFC3339，与 published_to 成对传入。
      schema:
        $ref: '#/components/schemas/UTCTimestamp'
    PublishedTo:
      name: published_to
      in: query
      required: false
      description: 发布时间半开区间上界，必须为 UTC RFC3339，与 published_from 成对传入。
      schema:
        $ref: '#/components/schemas/UTCTimestamp'
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 2048
...
    ThemeID:
      name: theme_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UUID'
...
  headers:
    RequestID:
      description: 本次 transport attempt 的 request id；重放时可以与首次请求不同。
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
  responses:
    BadRequest:
      description: 请求 DTO、query 或边界不合法
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: service bearer identity 缺失或无效
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: 已认证 principal 不具有 operation 所需 scope
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: caller scope 内不存在目标资源
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    UnprocessableEntity:
      description: 已解析 DTO 未通过 whole-batch 业务 validation；零 mutation
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    ResearchResourceLimit:
      description: 研究查询超过资源预算；整次失败且不返回静默截断结果
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResearchResourceLimitErrorEnvelope'
    ResearchAnalysisContextInconsistent:
      description: 实时页级引用无法闭合；调用方必须丢弃本轮结果并从第一页重查
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResearchAnalysisContextInconsistentErrorEnvelope'
    InternalError:
      description: Data Service 内部错误
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ServiceUnavailable:
      description: Event Semantics 依赖暂时不可用
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    UUID:
      type: string
      format: uuid
      description: RFC 4122 UUID string。
...
    UTCTimestamp:
      type: string
      format: date-time
      description: UTC RFC3339 timestamp；response 必须使用 UTC offset Z。
...
    ResearchThemeListEnvelope:
      type: object
      additionalProperties: false
      required: [request_id, result]
      properties:
        request_id: {type: string}
        result: {$ref: '#/components/schemas/ResearchThemeCollection'}
    ResearchThemeDetailEnvelope:
      type: object
      additionalProperties: false
      required: [request_id, result]
      properties:
        request_id: {type: string}
        result: {$ref: '#/components/schemas/ResearchThemeDetail'}
...
    ResearchAnalysisContextEnvelope:
      type: object
      additionalProperties: false
      required: [request_id, result]
      properties:
        request_id:
          type: string
        result:
          $ref: '#/components/schemas/ResearchAnalysisContext'
...